Track ranges in parser
This isn't perfect because I haven't done the work to get around
Parsec's tokeniser. When it parses

   x + y

the x token matches "x ". The fix isn't too bad - we just have to
write our own versions of the T.foo functions, but I'm going to punt
on that for now. The start position of stuff will be correct anyway.