Further changes to input language
This should now support crosses because we have an atomic "record this
set of things" defined by a group. (It's basically the same as a
covergroup).

The stuff in libdpi won't currently work (it's not updated to match)
and the reporting code isn't yet ported, so it won't even build.