Add a "cover bits" option
This doesn't allow you to cross a cover bits record with other
records, but it might be enough for now.

A note to self: If debugging this later, you might want something like
this command:

  (for x in $(cat run/ssim/system/.verilator-acov-logs.list); do \
     grep 'c:' $x | cut -d '{' -f 2 | tr -d ',}' | \
       tr ' ' '\n' | sed 's!^!0x!'; \
   done) | \
  xargs -n1 printf "%d\n" | \
  awk '{ acc = or(acc, compl($1)); } END { printf "%x\n", acc }'

This looks for record "c" (with the grep), then extracts the numbers,
printing one per line and prefixing them with 0x to make the base
clear. It then uses printf to convert to decimal and finally
accumulates all the bits with awk.

This version accumulates zeros. To accumulate ones, get rid of the
call to "compl".