Store "cover bits" records more efficiently
The previous DPI code stored every value seen into a text file. This
was then read in by acov-report and eventually mapped down to which
bits had been seen. That worked, but was hilariously inefficient if
there were lots of different values for the record.

The new code stores "cover bits" records differently. The DPI code
keeps a record of which bits have been one and which have been zero.
This will be way smaller than the list of all values that have been
seen.