Teach cover.cc to store narrow "cover bits" records more efficiently
Using "perf top", you can see that the remaining overhead is mostly in
the lookups in the various scope maps. We could get rid of this (~5%?)
by using vectors instead of maps. Before doing that, it would be good
to repack the indices so there are no gaps.

However, this patch seems worth doing: it speeds up a verilated test
here by a factor of two!