Fix broken dependencies in Makefile
This is trying to figure out whether we need to call cabal again to
rebuild. Cabal does the real dependency tracking, but we want Make to
know whether it needs to ask Cabal to do anything.

New GHCs places auto-generated .hi files in a different place, which
show up in hi-relpaths-for. Filtering with the set of files that
actually exist sorts it out.