Fix backwards reporting of cover bits
I thought the bug was probably because we were just passing a pair of
sets (ones and zeros) and that I'd swapped the order somewhere. So the
patch carefully names them (see brcZeros, brcOnes).

The eventual bug was in CountPass.hs in countGroup':

  ...
  where ...
        bc0 = mkBC w ones
        bc1 = mkBC w zeros

D'oh! I think the named version is probably better, so I'm making that
change too.